# gateway/docs/swagger.yaml
definitions:
  dto.AccountResponseDto:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      is_ban:
        type: boolean
      player_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - account_id
    - email
    - is_ban
    - player_id
    - username
    type: object
  dto.AccountsListResponseDto:
    properties:
      accounts:
        items:
          $ref: '#/definitions/dto.AccountResponseDto'
        type: array
      hasMore:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.ChangePassword:
    properties:
      account_id:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    required:
    - account_id
    - new_password
    - old_password
    type: object
  dto.LoginAccount:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginAccountResponse:
    properties:
      account_id:
        type: string
      player_id:
        type: string
      token:
        type: string
    required:
    - account_id
    - player_id
    - token
    type: object
  dto.RegisterAccount:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.RegisterAccountResponse:
    properties:
      account_id:
        type: string
      player_id:
        type: string
      token:
        type: string
    required:
    - account_id
    - player_id
    - token
    type: object
  dto.UpdateAccount:
    properties:
      account_id:
        type: string
      email:
        type: string
      username:
        type: string
    required:
    - account_id
    type: object
  dto.UpdateAccountResponse:
    properties:
      account_id:
        type: string
      updated_at:
        type: string
    required:
    - account_id
    type: object
info:
  contact:
    email: poomipat002@gmail.com
    name: Poomipat Chuealue
    url: https://github.com/Poomipat-Ch
paths:
  /account:
    put:
      consumes:
      - application/json
      description: 'Update account: username or email'
      parameters:
      - description: Update body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateAccountResponse'
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - Account
  /account/{id}:
    get:
      consumes:
      - application/json
      description: Get all information Account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountResponseDto'
      security:
      - BearerAuth: []
      summary: Account information by ID
      tags:
      - Account
  /account/password:
    put:
      consumes:
      - application/json
      description: Update password account
      parameters:
      - description: ChangePassword body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateAccountResponse'
      security:
      - BearerAuth: []
      summary: Change password account
      tags:
      - Account
  /account/search:
    get:
      consumes:
      - application/json
      description: Search Account with Username
      parameters:
      - description: Username
        in: query
        name: text
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountsListResponseDto'
      security:
      - BearerAuth: []
      summary: Search Account
      tags:
      - Account
  /login:
    post:
      consumes:
      - application/json
      description: Login with Username and Password
      parameters:
      - description: Login body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginAccountResponse'
      summary: Login
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Create new user account
      parameters:
      - description: Register body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterAccountResponse'
      summary: Register Account
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    in: header
    name: X-AUTH-TOKEN
    type: apiKey
swagger: "2.0"